Split ssd display from display view and add initial topology ssd
