+++ src/views/SchematicStatusDisplayView.vue
@@ -7,37 +7,21 @@
       :items="items"
     />
   </Teleport>
-  <div class="container">
-    <div
-      class="child-container"
-      :class="{ 'd-none': hideSSD }"
-      ref="ssdContainer"
-    >
-      <SsdComponent :src="src" @action="onAction" ref="ssdComponent" />
-      <DateTimeSlider
-        v-model:selectedDate="selectedDateSlider"
-        :dates="dates"
-        :hide-speed-controls="mobile"
-      />
-    </div>
-    <div class="child-container" :class="{ mobile, 'd-none': objectId === '' }">
-      <router-view @close="closeTimeSeriesDisplay"></router-view>
-    </div>
-  </div>
+  <SchematicStatusDisplay
+    :groupId="props.groupId"
+    :panelId="props.panelId"
+    :objectId="props.objectId"
+  />
 </template>
 
 <script setup lang="ts">
 import type {
-  SsdActionResult,
   SsdDisplayGroup,
   SsdDisplayPanel,
-  SsdActionRequest,
 } from '@deltares/fews-ssd-requests'
-import debounce from 'lodash-es/debounce'
 import { ref, onMounted, computed, watch } from 'vue'
 import { useRoute, useRouter } from 'vue-router'
 
-import { useAlertsStore } from '@/stores/alerts.ts'
 import { useUserSettingsStore } from '@/stores/userSettings.ts'
 
 import { configManager } from '../services/application-config/index.ts'
@@ -46,10 +30,7 @@ import type { ColumnItem } from '../components/general/ColumnItem'
 import { useSsd } from '../services/useSsd/index.ts'
 
 import HierarchicalMenu from '@/components/general/HierarchicalMenu.vue'
-import DateTimeSlider from '../components/general/DateTimeSlider.vue'
-import SsdComponent from '../components/ssd/SsdComponent.vue'
-import { useDisplay } from 'vuetify'
-import { useElementSize } from '@vueuse/core'
+import SchematicStatusDisplay from '@/components/ssd/SchematicStatusDisplay.vue'
 
 interface Props {
   groupId?: string
@@ -57,16 +38,7 @@ interface Props {
   objectId?: string
 }
 
-interface SsdActionEventPayload {
-  objectId: string
-  panelId: string
-  results: SsdActionResult[]
-}
-
-const sliderDebounceInterval = 500
-
 const baseUrl = configManager.get('VITE_FEWS_WEBSERVICES_URL')
-const alertsStore = useAlertsStore()
 const settings = useUserSettingsStore()
 const route = useRoute()
 const router = useRouter()
@@ -77,16 +49,10 @@ const props = withDefaults(defineProps<Props>(), {
   objectId: '',
 })
 
-const ssdComponent = ref<InstanceType<typeof SsdComponent> | null>(null)
-const ssdContainer = ref<HTMLElement | null>(null)
-
 const active = ref<string | undefined>(undefined)
 const open = ref<string[]>([])
 
 const selectedDate = ref<Date>(new Date())
-const selectedDateSlider = ref<Date>(selectedDate.value)
-
-const { mobile } = useDisplay()
 
 onMounted(() => {
   onGroupIdChange()
@@ -129,23 +95,11 @@ const items = computed(() => {
   return result
 })
 
-// Debounce the selected date from the slider input, so we do not send hundreds of requests when
-// dragging the slider around.
-watch(
-  selectedDateSlider,
-  debounce(
-    () => {
-      selectedDate.value = selectedDateSlider.value
-    },
-    sliderDebounceInterval,
-    { leading: true, trailing: true },
-  ),
-)
 // Make sure the appropriate group in the menu is open, and the panel is selected.
 watch(() => props.groupId, onGroupIdChange)
 watch(() => props.panelId, onPanelIdChange)
 
-const { capabilities, src, dates } = useSsd(
+const { capabilities } = useSsd(
   baseUrl,
   () => props.groupId,
   () => props.panelId,
@@ -191,143 +145,4 @@ function onGroupIdChange(): void {
 function onPanelIdChange(): void {
   active.value = props.panelId
 }
-
-const hideSSD = computed(() => {
-  return mobile.value && props.objectId !== ''
-})
-
-const ssdContainerSize = useElementSize(ssdContainer)
-watch(ssdContainerSize.width, () => {
-  if (ssdComponent.value) {
-    ssdComponent.value.resize()
-  }
-})
-
-function onAction(event: CustomEvent<SsdActionEventPayload>): void {
-  const { panelId, objectId, results } = event.detail
-  const now = new Date()
-  if (results.length === 0) {
-    alertsStore.addAlert({
-      id: `undefined-action-${now.toISOString()}`,
-      type: 'error',
-      message: 'No left click actions defined for this object',
-      active: true,
-    })
-    return
-  }
-
-  switch (results[0].type) {
-    case 'PDF':
-      window.open(new URL(results[0].requests[0].request))
-      break
-    case 'SSD':
-      switchPanel(results[0].requests[0])
-      break
-    case 'PI':
-      openTimeSeriesDisplay(panelId, objectId)
-      break
-    default:
-      alertsStore.addAlert({
-        id: `action-${results[0].type}-${now.toISOString()}`,
-        type: 'error',
-        message: `Action '${results[0].type}' not supported yet.`,
-        active: true,
-      })
-  }
-}
-
-function switchPanel(request: SsdActionRequest): void {
-  if (!capabilities.value) return
-
-  // We want to use the URL web API to parse the query parameters of the relative URL specified in
-  // the request; we are not actually using it as a URL. Hence, we use a random base URL.
-  const url = new URL(request.request, 'https://www.example.com')
-  const panelId = url.searchParams.get('ssd')
-
-  if (!panelId) return
-
-  // Find the display group that contains this panel.
-  const group = capabilities.value.displayGroups.find((cur) => {
-    return cur.displayPanels.some((panel) => panel.name === panelId)
-  })
-  const groupId = group?.name
-
-  if (!groupId) return
-
-  const currentRoute = router.currentRoute.value
-  const parentRoute = router
-    .getRoutes()
-    .find(
-      (route) =>
-        route.children &&
-        route.children.some((child) => child.name === currentRoute.name),
-    )
-  const targetRouteName = parentRoute?.name ?? currentRoute.name
-  if (!targetRouteName) return
-  router.push({
-    name: targetRouteName,
-    params: { groupId, panelId },
-    query: route.query,
-  })
-}
-
-function openTimeSeriesDisplay(panelId: string, objectId: string) {
-  const currentRoute = router.currentRoute.value
-  const routeConfig = router
-    .getRoutes()
-    .find((route) => route.name === currentRoute.name)
-  const childRoute = routeConfig?.children?.find((route) =>
-    route.name?.toString().endsWith('SSDTimeSeriesDisplay'),
-  )
-  router
-    .push({
-      name: childRoute?.name,
-      params: { objectId: objectId, panelId: panelId, groupId: props.groupId },
-    })
-    .then(() => {
-      ssdComponent.value?.resize()
-    })
-}
-
-function closeTimeSeriesDisplay(): void {
-  const currentRoute = router.currentRoute.value
-  const parentRoute = router
-    .getRoutes()
-    .find(
-      (route) =>
-        route.children &&
-        route.children.some((child) => child.name === currentRoute.name),
-    )
-  if (!parentRoute) return
-  router
-    .push({
-      name: parentRoute.name,
-      params: { groupId: props.groupId, panelId: props.panelId },
-    })
-    .then(() => {
-      ssdComponent.value?.resize()
-    })
-}
 </script>
-
-<style scoped>
-.container {
-  display: flex;
-  width: 100%;
-  height: 100%;
-}
-
-.child-container {
-  position: relative;
-  display: flex;
-  flex-direction: column;
-  width: 50%;
-  max-width: 100%;
-  flex: 1 1 0px;
-}
-
-.child-container.mobile {
-  height: 100%;
-  width: 100%;
-}
-</style>
